requirements.txt
flask
google-ads
